+++ packages/graph-explorer/src/connector/gremlin/fetchNeighborsCount/neighborsCountTemplate.test.ts
@@ -26,30 +26,4 @@ describe("Gremlin > neighborsCountTemplate", () => {
       `)
     );
   });
-
-  it("Should return a template for the given vertex id with defined limit", () => {
-    const template = neighborsCountTemplate({
-      vertexId: createVertexId("12"),
-      limit: 20,
-    });
-
-    expect(normalize(template)).toBe(
-      normalize(`
-        g.V("12").both().limit(20).dedup().group().by(label).by(count())
-      `)
-    );
-  });
-
-  it("Should return a template for the given vertex id with no limit", () => {
-    const template = neighborsCountTemplate({
-      vertexId: createVertexId("12"),
-      limit: 0,
-    });
-
-    expect(normalize(template)).toBe(
-      normalize(`
-        g.V("12").both().dedup().group().by(label).by(count())
-      `)
-    );
-  });
 });