+++ packages/graph-explorer/src/connector/sparql/sparqlExplorer.ts
@@ -177,7 +177,6 @@ export function createSparqlExplorer(
         _sparqlFetch(connection, featureFlags, options),
         {
           resourceURI: req.vertexId,
-          limit: req.limit,
         }
       );
     },